app1/requirements.txt
python-dotenv
flask
requests
azure-ai-inference
azure-core
gunicorn
sqlalchemy
sqlalchemy.orm
gunicorn